// types.tsx
export type RootStackParamList = {
  Root: undefined;
  NotFound: undefined;
};

export type BottomTabParamList = {
  GuestHome: undefined;
  TabTwo: undefined;
};

export type AllLinks = {
  GuestHome: undefined;
  LogIn: undefined;
  SignUp: undefined;
  UserHome: undefined;
  Help: undefined;
  Cart: undefined;
  Checkout: undefined;
  Delivery: undefined;
  DriverOrders: undefined;
  Payment: undefined;
};

export type TabTwoParamList = {
  TabTwoScreen: undefined;
};